New terrain for game world.
Please don’t edit the terrain. I know it’s not perfect, but it will at
least give us a frame of reference to work on. If you edit the position
of this terrain I honestly have no clue how much it will mess up the
rest of the game. So yea just don’t modify it.