Fixing the water interactions for the boat.
Created the buoyancy script which gives the boat floating like
properties. Also needed to create a specific mesh collider for the boat
so that script worked properly on the kayak form (needed to be a
completely closed off shape).

Added water effects to the boat with the mesh particle emitter
component (named it kayak spray). The spray could still use work, but
it’s working well enough for now.